+++ Source/JavaScriptCore/disassembler/UDis86Disassembler.cpp
@@ -49,7 +49,7 @@ bool tryToDisassembleWithUDis86(const MacroAssemblerCodePtr<DisassemblyPtrTag>& 
     uint64_t currentPC = disassembler.pc;
     while (ud_disassemble(&disassembler)) {
         char pcString[20];
-        snprintf(pcString, sizeof(pcString), "0x%lx", static_cast<unsigned long>(currentPC));
+        snprintf(pcString, sizeof(pcString), "0x%lx", static_cast<uintptr_t>(currentPC));
         out.printf("%s%16s: %s\n", prefix, pcString, ud_insn_asm(&disassembler));
         currentPC = disassembler.pc;
     }
